+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
+### 1.9.14-b1
+
+- **[FEATURE]** (`ktx-collections`) Added `getOrPut` extension function for `maps`
+
 #### 1.9.13-b1
 
 - **[UPDATE]** Updated to LibGDX 1.9.13.